VkOauth: WIP

+++ app/src/main/java/dev/ky3he4ik/chessproblems/presentation/view/users/UserListFragment.kt
@@ -1,22 +1,33 @@
 package dev.ky3he4ik.chessproblems.presentation.view.users
 
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import androidx.navigation.findNavController
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import dev.ky3he4ik.chessproblems.MainActivity
+import dev.ky3he4ik.chessproblems.R
 import dev.ky3he4ik.chessproblems.databinding.UserListFragmentBinding
+import dev.ky3he4ik.chessproblems.domain.model.users.UserInfo
+import dev.ky3he4ik.chessproblems.presentation.repository.Repository
+import dev.ky3he4ik.chessproblems.presentation.repository.model.users.UserInfoDTO
+import dev.ky3he4ik.chessproblems.presentation.view.problems.RecyclerItemClickListener
 import dev.ky3he4ik.chessproblems.presentation.view.users.adapters.UserListElementAdapter
 import dev.ky3he4ik.chessproblems.presentation.viewmodel.users.UserListViewModel
 
 class UserListFragment : Fragment() {
     private lateinit var viewModel: UserListViewModel
     private lateinit var binding: UserListFragmentBinding
+
+    private var isFabOpen = false
+    private var canDelete = false
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -34,16 +45,72 @@ class UserListFragment : Fragment() {
             ): Boolean = false
 
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                val position = viewHolder.adapterPosition
-                viewModel.deleteUser(
-                    (binding.userRecyclerView.adapter as UserListElementAdapter).data[position]
-                )
+                if (canDelete) {
+                    val position = viewHolder.adapterPosition
+                    viewModel.deleteUser(
+                        (binding.userRecyclerView.adapter as UserListElementAdapter).data[position]
+                    )
+                } else {
+                    Toast.makeText(context, "Access Denied", Toast.LENGTH_SHORT).show()
+                }
             }
         }).attachToRecyclerView(binding.userRecyclerView)
         binding.floatingActionButton2.setOnClickListener {
-            val action = UserListFragmentDirections.actionUserListToAddUser()
-            it.findNavController().navigate(action)
+            if (isFabOpen)
+                closeFabMenu()
+            else
+                openFabMenu()
+            isFabOpen = !isFabOpen
         }
+        binding.fabVk.setOnClickListener {
+            Repository.OAUTH.Vk.provider.auth { bundle ->
+                (activity as MainActivity).navHost.navController.navigate(
+                    R.id.action_userList_to_webFragment,
+                    bundle
+                )
+            }
+        }
+        binding.fabYandex.setOnClickListener {
+            Repository.OAUTH.Vk.provider.auth { bundle ->
+                (activity as MainActivity).navHost.navController.navigate(
+                    R.id.action_userList_to_webFragment,
+                    bundle
+                )
+            }
+        }
+        binding.fabGoogle.setOnClickListener {
+            Repository.OAUTH.Vk.provider.auth { bundle ->
+                (activity as MainActivity).navHost.navController.navigate(
+                    R.id.action_userList_to_webFragment,
+                    bundle
+                )
+            }
+        }
+
+        RecyclerItemClickListener.registerListener(context, binding.userRecyclerView,
+            object : RecyclerItemClickListener.OnItemClickListener {
+                override fun onItemClick(view: View?, position: Int) {
+                    val data =
+                        (binding.userRecyclerView.adapter as UserListElementAdapter).data
+                    if (position in data.indices) {
+                        Repository.activeUserId = data[position].userId
+                        setCanDelete()
+                        Toast.makeText(
+                            context,
+                            "Selected user: ${data[position].userId}",
+                            Toast.LENGTH_SHORT
+                        ).show()
+
+                    } else {
+                        Toast.makeText(context, "Invalid position: $position", Toast.LENGTH_SHORT)
+                            .show()
+                        Log.e("ProblemListFragment/click", "Invalid position: $position")
+                    }
+                }
+
+                override fun onLongItemClick(view: View?, position: Int) {}
+            })
+
         return binding.root
     }
 
@@ -54,5 +121,28 @@ class UserListFragment : Fragment() {
         viewModel.getUsersList().observe(viewLifecycleOwner, {
             binding.userRecyclerView.adapter = UserListElementAdapter(it)
         })
+        setCanDelete()
+    }
+
+    private fun openFabMenu() {
+        binding.fabGoogle.animate().translationY(-resources.getDimension(R.dimen.dp_210))
+        binding.fabYandex.animate().translationY(-resources.getDimension(R.dimen.dp_140))
+        binding.fabVk.animate().translationY(-resources.getDimension(R.dimen.dp_70))
+    }
+
+    private fun closeFabMenu() {
+        binding.fabGoogle.animate().translationY(0F)
+        binding.fabYandex.animate().translationY(0F)
+        binding.fabVk.animate().translationY(0F)
+    }
+
+    private fun setCanDelete() {
+        Repository.activeUserId?.let {
+            Repository.usersRepository.getUser<UserInfoDTO>(it)
+                .observe(viewLifecycleOwner, { user ->
+                    if (user != null)
+                        canDelete = user.roleLevel >= UserInfo.Roles.MODERATOR.roleLevel
+                })
+        }
     }
 }
